fix: remove log move nockRequest to file
